+++ header.php
@@ -2,11 +2,15 @@
 <html <?php language_attributes(); ?>>
 <head>
 <meta http-equiv="content-type" content="text/html; charset=<?php bloginfo('charset'); ?>">
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <meta name="description" content="">
 <meta name="keywords" content="">
 <meta name="author" content="">
-<?php wp_head(); ?>
+<?php if ( is_singular() && pings_open( get_queried_object() ) ) : ?>
+<link rel="pingback" href="<?php bloginfo( 'pingback_url' ); ?>">
+<?php endif;
+wp_head(); ?>
 </head>
 <?php $awada_theme_options = awada_theme_options();
 $class = "";