+++ home-portfolio.php
@@ -1,7 +1,7 @@
 <?php $awada_theme_options = awada_theme_options();
-$post_id = ($awada_theme_options['portfolio_post'] == ''?'':$awada_theme_options['portfolio_post']);
-$awada_portfolio_post = get_post($post_id); ?>
-<section id="home_portfolio" class="white-wrapper">
+// Theme_mod settings to check.
+$awada_portfolio_setting = $awada_theme_options[ 'awada_portfolio_setting' ]; ?>
+<section id="home_portfolio" class="white-wrapper sec-spacing">
 	<div class="container">
 		<div class="main_title">
 			<?php if($awada_theme_options['home_portfolio_title']!=""){ ?>
@@ -11,10 +11,31 @@ $awada_portfolio_post = get_post($post_id); ?>
 			<p id="portfolio_sub_heading" class="lead"><?php echo wp_kses_post($awada_theme_options['home_portfolio_subtitle']); ?></p>
 			<?php } ?>
 		</div><!-- end general title -->
-		<div class="padding-top"><?php
-			if($awada_portfolio_post->post_content!=''):
-				echo apply_filters('the_content',$awada_portfolio_post->post_content);
-			else:
+		<div class="padding-top">
+			<?php
+			if(!empty($awada_portfolio_setting) && is_array($awada_portfolio_setting)) {
+				foreach( $awada_portfolio_setting as $setting ) {
+					$post_id = $setting['content_page'];
+					$post = get_post( $post_id );
+					$post->post_excerpt;
+					$src = wp_get_attachment_image_src( get_post_thumbnail_id($post_id), 'awada_port_home_thumb' );
+					$url = $src[0]; ?>
+					<div class="col-lg-4 col-sm-4 col-md-4 col-sm-4">
+						<div class="portfolio_list">
+							<div class="content_entry">
+								<img src="<?php echo esc_url( $url ); ?>" alt="" class="img-responsive">
+								<div class="magnifier">
+									<div class="buttons">
+										<a class="st" rel="bookmark" href="<?php echo get_the_permalink($post_id); ?>"><i class="fa fa-link"></i></a>
+										<a class="sf" data-gal="prettyPhoto[product-gallery]" href="<?php echo esc_url( $url ); ?>"><i class="fa fa-expand"></i></a>
+										<h3><?php echo esc_html( $post->post_title ); ?></h3>
+									</div><!-- end buttons -->
+								</div><!-- end magnifier -->
+							</div><!-- end portfoliobox -->
+						</div><!-- end portfolio_list -->
+					</div><!-- end col-lg-4 -->
+				<?php }
+			} else {
 				$port_title = array(__('PERFECT THEME', 'awada'), __('QUICK SUPPORT', 'awada'), __('RESPONSIVE THEME', 'awada'), __('MULTIPLE COLOR SCHEME', 'awada'), __('RETINA READY', 'awada'), __('AMAZING SHORTCODES', 'awada'));
 				$j = 0;
 				for($i=1 ; $i<=3 ; $i++){ ?>
@@ -33,7 +54,7 @@ $awada_portfolio_post = get_post($post_id); ?>
 						</div><!-- end portfolio_list -->
 					</div><!-- end col-lg-4 -->
 				<?php $j++; }
-				endif; ?>	
+			} ?>	
 		</div>
 		<div class="clearfix"></div>
 	</div><!-- end container -->